+++ probe-rs/src/architecture/arm/sequences/nrf.rs
@@ -1,114 +1,14 @@
-//! Sequences for the nRF53.
-
-use std::sync::Arc;
+//! Sequences for the nRF devices.
 
 use super::ArmDebugSequence;
-use crate::architecture::arm::ap::{MemoryAp, CSW};
+use crate::architecture::arm::ap::MemoryAp;
 use crate::architecture::arm::{
     communication_interface::Initialized, ApAddress, ArmCommunicationInterface, ArmProbeInterface,
     DapAccess,
 };
 use crate::Memory;
 
-/// The sequence handle for the nRF5340.
-pub struct Nrf5340(());
-
-impl Nrf5340 {
-    /// Create a new sequence handle for the nRF5340.
-    pub fn create() -> Arc<dyn ArmDebugSequence> {
-        Arc::new(Self(()))
-    }
-}
-
-impl Nrf for Nrf5340 {
-    fn core_aps(&self, interface: &mut Memory) -> Vec<(ApAddress, ApAddress)> {
-        let ap_address = interface.get_ap();
-
-        let core_aps = [(0, 2), (1, 3)];
-
-        core_aps
-            .into_iter()
-            .map(|(core_ahb_ap, core_ctrl_ap)| {
-                (
-                    ApAddress {
-                        ap: core_ahb_ap,
-                        ..ap_address
-                    },
-                    ApAddress {
-                        ap: core_ctrl_ap,
-                        ..ap_address
-                    },
-                )
-            })
-            .collect()
-    }
-
-    fn is_core_unlocked(
-        &self,
-        arm_interface: &mut ArmCommunicationInterface<Initialized>,
-        ahb_ap_address: ApAddress,
-        _ctrl_ap_address: ApAddress,
-    ) -> Result<bool, crate::Error> {
-        let csw: CSW = arm_interface
-            .read_raw_ap_register(ahb_ap_address, 0x00)?
-            .into();
-        Ok(csw.DeviceEn != 0)
-    }
-
-    fn has_network_core(&self) -> bool {
-        true
-    }
-}
-
-/// The sequence handle for the nRF9160.
-pub struct Nrf9160(());
-
-impl Nrf9160 {
-    /// Create a new sequence handle for the nRF9160.
-    pub fn create() -> Arc<dyn ArmDebugSequence> {
-        Arc::new(Self(()))
-    }
-}
-
-impl Nrf for Nrf9160 {
-    fn core_aps(&self, interface: &mut Memory) -> Vec<(ApAddress, ApAddress)> {
-        let ap_address = interface.get_ap();
-
-        let core_aps = [(0, 4)];
-
-        core_aps
-            .into_iter()
-            .map(|(core_ahb_ap, core_ctrl_ap)| {
-                (
-                    ApAddress {
-                        ap: core_ahb_ap,
-                        ..ap_address
-                    },
-                    ApAddress {
-                        ap: core_ctrl_ap,
-                        ..ap_address
-                    },
-                )
-            })
-            .collect()
-    }
-
-    fn is_core_unlocked(
-        &self,
-        arm_interface: &mut ArmCommunicationInterface<Initialized>,
-        _ahb_ap_address: ApAddress,
-        ctrl_ap_address: ApAddress,
-    ) -> Result<bool, crate::Error> {
-        let approtect_status = arm_interface.read_raw_ap_register(ctrl_ap_address, 0x00C)?;
-        Ok(approtect_status != 0)
-    }
-
-    fn has_network_core(&self) -> bool {
-        false
-    }
-}
-
-trait Nrf: Sync + Send {
+pub trait Nrf: Sync + Send {
     /// Returns the ahb_ap and ctrl_ap of every core
     fn core_aps(&self, memory: &mut Memory) -> Vec<(ApAddress, ApAddress)>;