+++ CHANGELOG.md
@@ -5,9 +5,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[Added]
+### Added
 
 - Added PartialEq Trait to the struct DebugProbeInfo. (#1173)
+- Added support for configuring trace data destinations (#1177)
+
+### Changed
+- SWV vendor configuration has been refactored into sequences and trace functions have been renamed:
+    * `Session::setup_swv` has been renamed to `Session::setup_tracing`
+    * `Session::read_swo` has been renamed to `Session::read_trace_data`
+
+### Fixed
+* (#350) Flashing and debugging on STM32 chips using WFI instructions should now be stable (fixed in #1177)
 
 
 ## [0.13.0]